mpich-default: fix for PPC
requires libatomic, which means using gcc7 for build
not certain why it needs this on PPC only, as the build
succeeds with the standard gcc-4.2/apple-gcc4.2 on Intel.

It's possible that the test for libatomic might be tweaked to be more
accurrate, and then forcing gcc7 might not be needed. For now, using
gcc7 on PPC is not a hassle though.

And one file needed some trivial definitions added to satisfy some expected macros.